typescriptify pre-build
