+++ tsconfig.json
@@ -2,6 +2,8 @@
   "extends": "@tsconfig/docusaurus/tsconfig.json",
   "exclude": ["**/node_modules/*", "blog", "docs", "build", "i18n"],
   "compilerOptions": {
-    "baseUrl": "."
+    "baseUrl": ".",
+    "target": "ES2017",
+    "moduleResolution": "node"
   }
 }
